+++ Changelog.md
@@ -2,9 +2,9 @@
 
 ### Added
 
--   J-Link devices OB FW can be automatically update when Erasing, Writing and
-    Reading a JLink Device.
--   Button to Manually update OB Firmware of a device.
+-   J-Link devices OB FW can now be automatically updated when erasing, writing, and
+    reading.
+-   Button to manually update the OB Firmware of a device.
 
 ### Changed